Move codec tests into dedicated dev.zarr.zarrjava.codec package
Splits the codec-specific tests out of the monolithic ZarrV3Test into
one class per codec concern under a new codec/ test package:
TransposeCodecTest, ShardingCodecTest, BytesCodecTest,
CompressionCodecsTest (gzip/blosc/zstd) and CodecPipelineTest (invalid
codec ordering). Providers used only by moved tests move with them;
codecBuilders/chunkKeyEncodingsAndCodecs stay in ZarrV3Test since the
metadata-format regression test still uses them. Pure test reorg, no
behavior change.
